Do not pass context
