+++ app/src/main/java/com/mataku/scrobscrob/app/ui/LoginActivity.kt
@@ -13,6 +13,7 @@ import android.os.AsyncTask
 import android.os.Build
 import android.os.Bundle
 import android.provider.ContactsContract
+import android.provider.Settings
 import android.support.design.widget.Snackbar
 import android.support.v7.app.AppCompatActivity
 import android.text.TextUtils
@@ -62,7 +63,7 @@ class LoginActivity : AppCompatActivity(), LoaderCallbacks<Cursor>, LoginViewCal
 
         loginFormView = findViewById(R.id.login_form)
         progressView = findViewById(R.id.login_progress)
-        loginPresenter = LoginPresenter(applicationContext, this)
+        loginPresenter = LoginPresenter(isEnabledReadNotification(), this)
     }
 
     private fun populateAutoComplete() {
@@ -305,6 +306,21 @@ class LoginActivity : AppCompatActivity(), LoaderCallbacks<Cursor>, LoginViewCal
     private fun getSessionKey(): String {
         val data = getSharedPreferences("DATA", Context.MODE_PRIVATE)
         return data.getString("SessionKey", "")
+    }
+
+    private fun isEnabledReadNotification(): Boolean {
+        val contentResolver = contentResolver
+        val rawListeners = Settings.Secure.getString(contentResolver,
+                "enabled_notification_listeners")
+        if (rawListeners == null || rawListeners == "") {
+            return false
+        } else {
+            val listeners = rawListeners.split(":".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()
+            listeners
+                    .filter { it.startsWith(packageName) }
+                    .forEach { return true }
+        }
+        return false
     }
 
     companion object {
